subreddit modlist unimplemented. Moved parse-json to its own file
