+++ src/demo.js
@@ -1,11 +1,12 @@
-import { RedditClient, getPostsFromSubreddit} from './RedditClient.js';
+import { RedditClient, getPostsFromSubreddit, getSubredditModList} from './RedditClient.js';
 
 
 
 const redditClient = new RedditClient();
 
-const commentsList = await getPostsFromSubreddit(1, 'test', 'new');
+// const commentsList = await getPostsFromSubreddit(1, 'test', 'new');
+const modList = await getSubredditModList('test');
 
 // const commentsList = await redditClient.getLatestCommentsFromReddit(1);
 
-console.log(commentsList);
+console.log(modList);
